+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fix documentation for --allow-unsafe-core-replacement #4341
 - fix missing linebreaks in quotes #4360
 - avoid showing wrong menu items for blocked users #4353
+- fix: save message draft every 200ms if message text changed #3733
 
 <a id="1_48_0"></a>